requirements.txt
svg.path==3.0
